added view by category
added functionality for clicking on the category listed with each post,
which displays a potentially paginated feed of the articles in a
particular category. accomplished primarily via adding methods to the
BlogRepository and ListViewModel to generate a model in much the same
way as for displaying all posts except filtered by category.